+++ playground/vanilla/app/serializers/composer_serializer.rb
@@ -1,11 +1,8 @@
 class ComposerSerializer < BaseSerializer
   attributes(
     :id,
-    :first_name,
     :last_name,
-  )
-
-  typed_attributes(
-    name: :string,
+    :first_name,
+    name: {type: :string},
   )
 end