+++ playground/vanilla/app/serializers/snake_composer_serializer.rb
@@ -1,0 +1,3 @@
+class SnakeComposerSerializer < ComposerSerializer
+  transform_keys ->(key) { key }
+end