+++ spec/types_from_serializers/generator_spec.rb
@@ -6,6 +6,7 @@ describe "Generator" do
   let(:sample_dir) { Rails.root.join("app/frontend/types/serializers") }
   let(:serializers) {
     %w[
+      Nested::AlbumSerializer
       VideoWithSongSerializer
       VideoSerializer
       SongSerializer
@@ -13,11 +14,12 @@ describe "Generator" do
       ModelSerializer
       ComposerWithSongsSerializer
       ComposerSerializer
+      SnakeComposerSerializer
     ]
   }
 
   def file_for(dir, name)
-    dir.join("#{name.chomp("Serializer")}.ts")
+    dir.join("#{name.chomp("Serializer").gsub("::", "/")}.ts")
   end
 
   def app_file_for(name)
@@ -60,13 +62,13 @@ describe "Generator" do
     # It generates one file per serializer.
     serializers.each do |name|
       output_file = output_file_for(name)
-      expect(output_file.read).to match_snapshot("interfaces_#{name}")
+      expect(output_file.read).to match_snapshot("interfaces_#{name}") # UPDATE_SNAPSHOTS="1" bin/rspec
     end
 
     # It generates an file that exports all interfaces.
     index_file = output_dir.join("index.ts")
     expect(index_file.exist?).to be true
-    expect(index_file.read).to match_snapshot("interfaces_index")
+    expect(index_file.read).to match_snapshot("interfaces_index") # UPDATE_SNAPSHOTS="1" bin/rspec
 
     # It does not render if generating again.
     TypesFromSerializers.generate